requirements.txt
streamlit>=1.8.1
streamlit-lottie==0.0.3
requests==2.27.1
Pyrebase4==4.5.0
scikit-learn==1.0.2
pandas>=1.0.5
numpy==1.16.5
xlrd >= 1.2.0
openpyxl==3.0.9
protobuf==3.20.1
